+++ src/lang/isNaN.js
@@ -7,7 +7,7 @@ define(['./isNumber'], function (isNumber) {
         // based on the fact that NaN !== NaN
         // need to check if it's a number to avoid conflicts with host objects
         // also need to coerce ToNumber to avoid edge case `new Number(NaN)`
-        return isNumber(val) && val != +val;
+        return !isNumber(val) || val != +val;
     }
 
     return isNaN;